chore: script

+++ tools/update_compact_dicts.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 
+STRICT=x"$1"
+
+echo Strict about errors? $STRICT
+
 compact_dicts=(
     "../moran.essay.dict.yaml"
     "../moran.tencent.dict.yaml"
-    "../moran.genshin.dict.yaml"
     "../moran.moe.dict.yaml"
     "../moran.thuocl.dict.yaml"
 )
@@ -31,7 +34,7 @@ update_compact_dict() {
     extract_dict "$DICT_FILE" "$HEADER_FILE" "$INPUT_FILE"
     python3 schemagen.py update-compact-dict --rime-dict="$INPUT_FILE" > "$OUTPUT_FILE"
 
-    if grep '^# BAD' "$OUTPUT_FILE"
+    if [ $STRICT = x"yes" ] && grep '^# BAD' "$OUTPUT_FILE"
     then
         echo '!!! BAD DICT !!!'
         rm -f $INPUT_FILE $HEADER_FILE
